# src/main/resources/static/openapi.yml
openapi: 3.1.0
info:
  title: Computer Architecture Course Administration API
  description: This API provides backend functionality for managing the Computer Architecture course system.
#  termsOfService: ?
#  contact:
#    name: ?
#    email: ?
#    url: ?
#  license:
#    name: ?
#    url: ?
  version: 1.0.0
#externalDocs:
#  description: ?
#  url: ?
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/exercise:
    get:
      tags:
        - ExerciseController
      summary: Retrieve a list of exercises
      description: >
        This endpoint allows you to fetch a list of exercises with optional sorting and filtering.

        - **Sorting** is available for all fields defined in `ExerciseRequestDTO` (including `id`).

        - **Filtering** is available for `id` (can be used at the same time).
      operationId: getExercisesByCriteria
      parameters:
        - $ref: '#/components/parameters/sort'
        - name: id
          in: query
          description: Filter exercises by ID.
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK - List of exercises returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExerciseResponseDTO'
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    post:
      tags:
        - ExerciseController
      summary: Create a new exercise/exercises
      description: Adds a new exercise/exercises to the system with the provided details.
      operationId: addExercises
      requestBody:
        description: Array of exercise objects that need to be created.
        content:
          application/json:
            schema:
              type: object
              properties:
                exercises:
                  type: array
                  items:
                    $ref: '#/components/schemas/ExerciseRequestDTO'
        required: true
      responses:
        '201':
          description: Created - Exercises created successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SuccessResponseDTO'
        '400':
          description: Bad Request - Required request body is missing - Bad JSON format.
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        '422':
          description: Unprocessable Entity - Validation Failed.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    put:
      tags:
        - ExerciseController
      summary: Update exercise by ID
      description: Updates information for an existing exercise identified by its ID.
      operationId: updateExerciseById
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the exercise.
          schema:
            type: integer
            format: int32
      requestBody:
        description: Exercise object with updated fields.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExerciseRequestDTO'
        required: true
      responses:
        '200':
          description: OK - Exercise updated successfully.
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/SuccessResponseDTO'
        '400':
          description: Bad Request - Required request body is missing - Bad JSON format.
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '404':
          description: Not Found - Exercise not found.
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        '422':
          description: Unprocessable Entity - Validation Failed.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    delete:
      tags:
        - ExerciseController
      summary: Delete exercise by ID
      description: Removes a exercise from the system identified by its ID.
      operationId: deleteExerciseById
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the exercise.
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK - Exercise deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDTO'
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '404':
          description: Not Found - Exercise not found.
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
...
  /api/v1/student-assignment:
    get:
      tags:
        - StudentAssignmentController
      summary: Retrieve a list of studentAssignments
      description: >
        This endpoint allows you to fetch a list of studetAssignments with optional sorting and filtering.

        - **Sorting** is available for all fields defined in `StudentAssignmentRequestDTO` (including `id`).

        - **Filtering** is available for `id` (can be used at the same time).
      operationId: getStudentAssignmentsByCriteria
      parameters:
        - $ref: '#/components/parameters/sort'
        - name: id
          in: query
          description: Filter studentAssignments by ID.
          required: false
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK - List of studentAssignments returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentAssignmentResponseDTO'
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
    post:
      tags:
        - StudentAssignmentController
      summary: Create a new studentAssignment/studentAssignments
      description: Adds a new studentAssignment/studentAssignments to the system with the provided details.
      operationId: addStudentAssignments
      requestBody:
        description: Array of studentAssignment objects that need to be created.
        content:
          application/json:
            schema:
              type: object
              properties:
                studentAssignments:
                  type: array
                  items:
                    $ref: '#/components/schemas/StudentAssignmentRequestDTO'
        required: true
      responses:
        '201':
          description: Created - StudentAssignments created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDTO'
        '400':
          description: Bad Request - Required request body is missing - Bad JSON format.
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '404':
          description: Not Found - Assignment, Student or ExerciseSession not found.
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        '422':
          description: Unprocessable Entity - Validation Failed.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    put:
      tags:
        - StudentAssignmentController
      summary: Update studentAssignment by ID
      description: Updates information for an existing studentAssignment identified by its ID.
      operationId: updateStudentAssignmentById
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the studentAssignment.
          schema:
            type: integer
            format: int32
      requestBody:
        description: StudentAssignment object with updated fields.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentAssignmentRequestDTO'
        required: true
      responses:
        '200':
          description: OK - StudentAssignment updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDTO'
        '400':
          description: Bad Request - Required request body is missing - Bad JSON format.
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '404':
          description: Not Found - StudentAssignment, Assignment, Student or ExerciseSession not found.
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        '422':
          description: Unprocessable Entity - Validation Failed.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
    delete:
      tags:
        - StudentAssignmentController
      summary: Delete studentAssignment by ID
      description: Removes a studentAssignment from the system identified by its ID.
      operationId: deleteStudentAssignmentById
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the studentAssignment.
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK - StudentAssignment deleted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponseDTO'
        '401':
          description: "Unauthorised - Unauthenticated: Authentication is required to access this resource."
        '403':
          description: "Forbidden - Access denied: You do not have permission to perform this action."
        '404':
          description: Not Found - StudentAssignment not found.
        '405':
          description: Method Not Allowed - Incorrect HTTP method used.
        '409':
          description: Conflict - The request could not be processed due to data integrity rules violation.
        default:
          description: Unexpected error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDTO'
...
components:
  parameters:
    sort:
      name: sort
      in: query
      required: false
      description: >
        Sort the results by none or any field from the corresponding `Request/ResponseDTO`.
        Format: `field,(asc|desc)` >>> for example: `id,asc` or `id,desc`.
        If no order is specified, default is ascending.  
        Multiple sort parameters can be included, like:  
        `sort=field1,asc&sort=field2,desc`
      schema:
        type: string
  schemas:
...
    ErrorResponseDTO:
      type: object
      description: Represents a standard error response containing status code, general message, and optional list of detailed error messages.
      required:
        - status
        - message
      properties:
        status:
          type: integer
          format: int32
          description: HTTP status code representing the error.
        message:
          type: string
          description: A short description or summary of the error.
        errors:
          type: array
          items:
            type: string
          description: A list of detailed error messages.
...
    ExerciseRequestDTO:
      type: object
      description: Request body used for creating new exercise or updating existing exercise.
      required:
        - firstSessionDate
        - startTime
        - roomEnum
      properties:
        firstSessionDate:
          type: LocalDate
          description: Date when first session of this exercise starts.
        startTime:
          type: string
          format: time
          description: Start time of the exercise.
        roomEnum:
          type: string
          description: The room where the exercise takes place.
...
    StudentAssignmentRequestDTO:
      type: object
      description: Request body used for creating new studentAssignment or updating existing studentAssignment.
      required:
        - assignmentId
        - studentId
        - earnedPoints
      properties:
        assignmentId:
          type: integer
          format: int32
          description: The unique identifier for the assignment associated with the studentAssignment.
        studentId:
          type: integer
          format: int32
          description: The unique identifier for the student associated with the studentAssignment.
        note:
          type: string
          description: Optional reason why certain amount of points was assigned.
        earnedPoints:
          type: number
          format: double
          description: The number of points the student earned for the given assignment.
...
    SuccessResponseDTO:
      type: object
      description: Represents a standard success response containing status code, general message, and optional list of data.
      required:
        - status
        - message
      properties:
        status:
          type: integer
          format: int32
          description: HTTP status code representing the success.
        message:
          type: string
          description: A short description or summary of the success.
        data:
          type: array
          items:
            type: object
          description: A list of updated or deleted data.
...
    ExerciseResponseDTO:
      type: object
      description: Response body that returns details about an exercise.
      required:
        - id
        - firstSessionDate
        - startTime
        - roomEnum
      properties:
        id:
          type: integer
          format: int32
          description: The exercise's unique identifier.
        firstSessionDate:
          type: LocalDate
          description: Date when first session of this exercise starts.
        startTime:
          type: string
          format: time
          description: The starting time of the exercise.
        roomEnum:
          type: string
          description: The room where the exercise takes place.
...
    StudentAssignmentResponseDTO:
      type: object
      description: Response body that returns details about a studentAssignment.
      required:
        - id
        - assignment
        - student
        - earnedPoints
        - createdBy
        - createdAt
        - updatedBy
        - updatedAt
      properties:
        id:
          type: integer
          format: int32
          description: The studentAssignment's id.
        assignment:
          type: object
          description: Details about the assignment associated with the studentAssignment.
        student:
          type: object
          description: Basic information about the student to whom the assignment is assigned.
        note:
          type: string
          description: Optional reason why certain amount of points was assigned.
        earnedPoints:
          type: number
          format: double
          description: The number of points the student earned for the given assignment.
        createdBy:
          type: object
          format: int32
          description: Information about the user who created the studentAssignment.
        createdAt:
          type: string
          format: date-time
          description: The timestamp when the studentAssignment was created.
        updatedBy:
          type: integer
          format: int32
          description: Information about the user who last updated the studentAssignment.
        updatedAt:
          type: string
          format: date-time
          description: The timestamp when the studentAssignment was last updated.
...